pyproject.toml
[tool.poetry]
name = "mini-django-app"
version = "0.0.1"
description = ""
authors = ["Anyone Person"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
orjson = "3.10.4"
pydantic-settings = "^2.2.1"
sentry-sdk = "^2.3.1"
uvloop = "^0.19.0"
django = "^5.0.6"
django-ninja = "^1.1.0"
celery = "^5.4.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
django-cors-headers = "^4.3.1"
numpy = "^2.0.0"
redis = "^5.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-asyncio = "^0.23.7"
ruff = "^0.4.5"
mypy = "^1.10.0"
django-stubs = "^5.0.2"
poetry-plugin-export = "^1.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = [
    ".git",
    "__pycache__",
    ".venv/*",
    "*.pyc",
    ".mypy_cache",
    ".pytest_cache",
    "setup.py",
]
line-length = 120
indent-width = 4
force-exclude = true
show-fixes = true
output-format = "concise"
target-version = "py312"

[tool.ruff.lint]
select = [
    "I", "ANN", "D", "F", "E", "W", "N", "C90", "PL",
    "T20", "SLF", "SIM", "TID", "TCH", "PERF", "FURB"
]
ignore = [
    "D100", "D104", "D107", "D203", "D212", "D401", "F403",
    "N805", "ANN101", "ANN102", "ANN401", "PLR", "PLW0603",
]
per-file-ignores = { "tests/*" = ["ANN001", "ANN201", "D103", "PLC2701"] }
fixable = ["ALL"]
unfixable = []
dummy-variable-rgx = "(?x)(^venv/ | ^venv\\.py$)"

[tool.ruff.format]
line-ending = "auto"
indent-style = "space"
quote-style = "double"
skip-magic-trailing-comma = false
docstring-code-format = true
docstring-code-line-length = 40

[tool.mypy]
strict = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
strict_equality = true
follow_imports = "silent"
exclude = 'tests/.*'
